refactor: carry agent-task request opaquely in runner execution envelope
runner_execution_envelope (a core type consumed by cli + lab-runner, not
agent) embedded AgentTaskRequest directly: the `agent_task` field was typed
Option<AgentTaskRequest>, plus from_agent_task_request() and
From<AgentTaskArtifactDeclaration> constructors — a core->agent-task upward
edge in runner_execution_envelope.rs.

The field is never read as a typed AgentTaskRequest outside the module, and
the two constructors had zero production callers (test-only). Make the field
Option<serde_json::Value> (carried opaquely; byte-identical wire format) and
move the request->envelope conversion to the agent side as
AgentTaskRequest::to_runner_execution_envelope(), along with its two tests.
Core keeps the envelope type; the agent-task layer owns the conversion and
opaque deserialization.

Stage 2 prep-PR of the homeboy-agents extraction campaign. Severs
runner_execution_envelope.rs's agent-task dependency (edge files 15 -> 14).
